* releases: (464 commits)
  COMP: Fix typo in first line
  Merge Scripts and antsSCCANObject.hxx from master commit ea9a0e7b2f088e24976fc1c253ecef791a2643bd
  BUG: set spacing in time series image
  COMP: vcl_pow
  ENH: v2.0.1. release
  BUG: review name wrong
  WIP: release prep
  Update README.md
  ENH: inverse warped to quick syn
  DOC: more notes on tx direction.
  Update README.md
  Update README.md
  ENH:  Adding voxelwise max.
  WIP: ...
  WIP: meaning of covering in cca different than pca
  WIP: PM smoothing; BUG: convergence of binary search
  WIP: Perona-Malik time step
  WIP: avoid warning in PM filtering
  WIP: Perona-Malik time step
  ENH: Perona-Malik in ImageMath
  ...

+++ Scripts/antsAtroposN4.sh
@@ -30,7 +30,7 @@ Usage:
 
 Example:
 
-  bash $0 -d 3 -i t1.nii.gz -x mask.nii.gz -l segmentationTemplate.nii.gz -p segmentationPriors%d.nii.gz -o output
+  bash $0 -d 3 -a t1.nii.gz -x mask.nii.gz -c 4 -p segmentationPriors%d.nii.gz -o output
 
 Required arguments:
 
@@ -38,13 +38,8 @@ Required arguments:
      -a:  input image                           Anatomical image, typically T1.  If more than one
                                                 anatomical image is specified, subsequently specified
                                                 images are used during the segmentation process.
-     -x:  mask image                            Mask defining the region of interest.
-     -m:  max. N4 <-> Atropos iterations        Maximum number of (outer loop) iterations between N4 <-> Atropos.
-     -n:  max. Atropos iterations               Maximum number of (inner loop) iterations in Atropos.
+     -x:  mask image                            Binary mask defining the region of interest.
      -c:  number of segmentation classes        Number of classes defining the segmentation
-     -l:  posterior label for N4 weight mask    Which posterior probability image should be used to define the
-                                                N4 weight mask.  Can also specify multiple posteriors in which
-                                                case the chosen posteriors are added.
      -o:  output prefix                         The following images are created:
                                                   * ${OUTPUT_PREFIX}N4Corrected.${OUTPUT_SUFFIX}
                                                   * ${OUTPUT_PREFIX}Segmentation.${OUTPUT_SUFFIX}
@@ -52,14 +47,35 @@ Required arguments:
 
 Optional arguments:
 
+     -m:  max. N4 <-> Atropos iterations        Maximum number of (outer loop) iterations between N4 <-> Atropos.
+     -n:  max. Atropos iterations               Maximum number of (inner loop) iterations in Atropos.
      -p:  segmentation priors                   Prior probability images initializing the segmentation.
                                                 Specified using c-style formatting, e.g. -p labelsPriors%02d.nii.gz.
      -r:  mrf                                   Specifies MRF prior (of the form '[weight,neighborhood]', e.g.
                                                 '[0.1,1x1x1]' which is default).
-     -b:  posterior formulation                 'Posterior formulation and whether or not to use mixture model proportions.'
+     -b:  posterior formulation                 Posterior formulation and whether or not to use mixture model proportions.
+                                                e.g 'Socrates[1]' (default) or 'Aristotle[1]'.  Choose the latter if you
+                                                want use the distance priors (see also the -l option for label propagation
+                                                control).
+     -l:  label propagation                     Incorporate a distance prior one the posterior formulation.  Should be
+                                                of the form 'label[lambda,boundaryProbability]' where label is a value
+                                                of 1,2,3,... denoting label ID.  The label probability for anything
+                                                outside the current label
+
+                                                  = boundaryProbability * exp( -lambda * distanceFromBoundary )
+
+                                                Intuitively, smaller lambda values will increase the spatial capture
+                                                range of the distance prior.  To apply to all label values, simply omit
+                                                specifying the label, i.e. -l [lambda,boundaryProbability].
+     -y:  posterior label for N4 weight mask    Which posterior probability image should be used to define the
+                                                N4 weight mask.  Can also specify multiple posteriors in which
+                                                case the chosen posteriors are combined.
      -s:  image file suffix                     Any of the standard ITK IO formats e.g. nrrd, nii.gz (default), mhd
-     -k:  keep temporary files                  Keep temporary files on disk (default = false).
+     -k:  keep temporary files                  Keep temporary files on disk (default = 0).
+     -u:  use random seeding                    Use random number generated from system clock in Atropos (default = 1)
      -w:  Atropos prior segmentation weight     Atropos spatial prior probability weight for the segmentation (default = 0)
+
+     -z:  Test / debug mode                     If > 0, attempts to continue after errors.
 
 USAGE
     exit 1
@@ -88,24 +104,44 @@ echoParameters() {
        posterior formulation  = ${ATROPOS_SEGMENTATION_POSTERIOR_FORMULATION}
        mrf                    = ${ATROPOS_SEGMENTATION_MRF}
        Max N4->Atropos iters. = ${ATROPOS_SEGMENTATION_NUMBER_OF_ITERATIONS}
+       use clock random seed  = ${USE_RANDOM_SEEDING}
 
 PARAMETERS
 }
 
 
-#    local  myresult='some value'
-#    echo "$myresult"
+# Echos a command to stdout, then runs it
+# Will immediately exit on error unless you set debug flag
+DEBUG_MODE=0
 
-# Echos a command to both stdout and stderr, then runs it
 function logCmd() {
   cmd="$*"
   echo "BEGIN >>>>>>>>>>>>>>>>>>>>"
   echo $cmd
-  logCmdOutput=$( $cmd | tee /dev/tty )
+
+  exec 5>&1
+  logCmdOutput=$( $cmd | tee >(cat - >&5) )
+
+  cmdExit=${PIPESTATUS[0]}
+
+  if [[ $cmdExit -gt 0 ]];
+    then
+      echo "ERROR: command exited with nonzero status $cmdExit"
+      echo "Command: $cmd"
+      echo
+      if [[ ! $DEBUG_MODE -gt 0 ]];
+        then
+          exit 1
+        fi
+    fi
+
   echo "END   <<<<<<<<<<<<<<<<<<<<"
   echo
   echo
+
+  return $cmdExit
 }
+
 
 ################################################################################
 #
@@ -122,6 +158,8 @@ OUTPUT_PREFIX=${OUTPUT_DIR}/tmp
 OUTPUT_SUFFIX="nii.gz"
 
 KEEP_TMP_IMAGES=0
+
+USE_RANDOM_SEEDING=1
 
 DIMENSION=3
 
@@ -149,15 +187,15 @@ ATROPOS_SEGMENTATION_LIKELIHOOD="Gaussian"
 ATROPOS_SEGMENTATION_POSTERIOR_FORMULATION="Socrates[1]"
 ATROPOS_SEGMENTATION_MASK=''
 ATROPOS_SEGMENTATION_NUMBER_OF_ITERATIONS=5
-ATROPOS_SEGMENTATION_NUMBER_OF_ITERATIONS=5
 ATROPOS_SEGMENTATION_NUMBER_OF_CLASSES=3
 ATROPOS_SEGMENTATION_MRF=''
+ATROPOS_SEGMENTATION_LABEL_PROPAGATION=()
 
 if [[ $# -lt 3 ]] ; then
   Usage >&2
   exit 1
 else
-  while getopts "a:b:c:d:h:k:l:m:n:o:p:r:s:t:w:x:" OPT
+  while getopts "a:b:c:d:h:k:l:m:n:o:p:r:s:t:u:w:x:y:z:" OPT
     do
       case $OPT in
           c) #number of segmentation classes
@@ -184,8 +222,8 @@ else
           k) #keep tmp images
        KEEP_TMP_IMAGES=$OPTARG
        ;;
-          l) #
-       N4_WEIGHT_MASK_POSTERIOR_LABELS[${#N4_WEIGHT_MASK_POSTERIOR_LABELS[@]}]=$OPTARG
+          l)
+       ATROPOS_SEGMENTATION_LABEL_PROPAGATION[${#ATROPOS_SEGMENTATION_LABEL_PROPAGATION[@]}]=$OPTARG
        ;;
           m) #atropos segmentation iterations
        N4_ATROPOS_NUMBER_OF_ITERATIONS=$OPTARG
@@ -208,11 +246,20 @@ else
           t) #n4 convergence
        N4_CONVERGENCE=$OPTARG
        ;;
+          u) #use random seeding
+       USE_RANDOM_SEEDING=$OPTARG
+       ;;
           w) #atropos prior weight
        ATROPOS_SEGMENTATION_PRIOR_WEIGHT=$OPTARG
        ;;
           x) #atropos segmentation mask
        ATROPOS_SEGMENTATION_MASK=$OPTARG
+       ;;
+          y) #
+       N4_WEIGHT_MASK_POSTERIOR_LABELS[${#N4_WEIGHT_MASK_POSTERIOR_LABELS[@]}]=$OPTARG
+       ;;
+          z) #debug mode
+       DEBUG_MODE=$OPTARG
        ;;
           *) # getopts issues an error message
        echo "ERROR:  unrecognized option -$OPT $OPTARG"
@@ -361,8 +408,10 @@ if [[ $INITIALIZE_WITH_KMEANS -eq 0 ]]
         N4_WEIGHT_MASK_IMAGES=( ${N4_WEIGHT_MASK_IMAGES[@]} ${PRIOR_IMAGE_FILENAMES[${N4_WEIGHT_MASK_POSTERIOR_IDXS[$i]}]} )
       done
 
-    logCmd ${ANTSPATH}ImageMath ${DIMENSION} ${SEGMENTATION_WEIGHT_MASK} PureTissueN4WeightMask ${N4_WEIGHT_MASK_IMAGES[@]}
-
+    if [[ ${#N4_WEIGHT_MASK_IMAGES[@]} -gt 0 ]];
+      then
+        logCmd ${ANTSPATH}ImageMath ${DIMENSION} ${SEGMENTATION_WEIGHT_MASK} PureTissueN4WeightMask ${N4_WEIGHT_MASK_IMAGES[@]}
+      fi
   fi
 
 if [[ -f ${SEGMENTATION_CONVERGENCE_FILE} ]];
@@ -377,9 +426,12 @@ for (( i = 0; i < ${N4_ATROPOS_NUMBER_OF_ITERATIONS}; i++ ))
     for(( j = 0; j < ${#ANATOMICAL_IMAGES[@]}; j++ ))
       do
         SEGMENTATION_N4_IMAGES=( ${SEGMENTATION_N4_IMAGES[@]} ${ATROPOS_SEGMENTATION_OUTPUT}${j}N4.${OUTPUT_SUFFIX} )
-
-        logCmd ${ANTSPATH}ImageMath ${DIMENSION} ${SEGMENTATION_N4_IMAGES[$j]} TruncateImageIntensity ${ANATOMICAL_IMAGES[$j]} 0.025 0.995 256 ${ATROPOS_SEGMENTATION_MASK}
-
+        if [[ $j == 0 ]];
+          then
+            logCmd ${ANTSPATH}ImageMath ${DIMENSION} ${SEGMENTATION_N4_IMAGES[$j]} TruncateImageIntensity ${ANATOMICAL_IMAGES[$j]} 0.025 0.995 256 ${ATROPOS_SEGMENTATION_MASK}
+          else
+            cp ${ANATOMICAL_IMAGES[$j]} ${SEGMENTATION_N4_IMAGES[$j]}
+          fi
         exe_n4_correction="${N4} -d ${DIMENSION} -i ${SEGMENTATION_N4_IMAGES[$j]} -x ${ATROPOS_SEGMENTATION_MASK} -s ${N4_SHRINK_FACTOR} -c ${N4_CONVERGENCE} -b ${N4_BSPLINE_PARAMS} -o ${SEGMENTATION_N4_IMAGES[$j]}"
         if [[ -f ${SEGMENTATION_WEIGHT_MASK} ]];
           then
@@ -407,8 +459,14 @@ for (( i = 0; i < ${N4_ATROPOS_NUMBER_OF_ITERATIONS}; i++ ))
           fi
       fi
 
-    exe_segmentation="${ATROPOS} -d ${DIMENSION} -x ${ATROPOS_SEGMENTATION_MASK} -c ${ATROPOS_SEGMENTATION_CONVERGENCE} ${ATROPOS_ANATOMICAL_IMAGES_COMMAND_LINE}"
-    exe_segmentation="${exe_segmentation} -i ${INITIALIZATION} -k ${ATROPOS_SEGMENTATION_LIKELIHOOD} -m ${ATROPOS_SEGMENTATION_MRF} -o [${ATROPOS_SEGMENTATION},${ATROPOS_SEGMENTATION_POSTERIORS}]"
+    ATROPOS_LABEL_PROPAGATION_COMMAND_LINE=''
+    for (( j = 0; j < ${#ATROPOS_SEGMENTATION_LABEL_PROPAGATION[@]}; j++ ))
+      do
+        ATROPOS_LABEL_PROPAGATION_COMMAND_LINE="${ATROPOS_LABEL_PROPAGATION_COMMAND_LINE} -l ${ATROPOS_SEGMENTATION_LABEL_PROPAGATION[$j]}";
+      done
+
+    exe_segmentation="${ATROPOS} -d ${DIMENSION} -x ${ATROPOS_SEGMENTATION_MASK} -c ${ATROPOS_SEGMENTATION_CONVERGENCE} ${ATROPOS_ANATOMICAL_IMAGES_COMMAND_LINE} ${ATROPOS_LABEL_PROPAGATION_COMMAND_LINE}"
+    exe_segmentation="${exe_segmentation} -i ${INITIALIZATION} -k ${ATROPOS_SEGMENTATION_LIKELIHOOD} -m ${ATROPOS_SEGMENTATION_MRF} -o [${ATROPOS_SEGMENTATION},${ATROPOS_SEGMENTATION_POSTERIORS}] -r ${USE_RANDOM_SEEDING}"
 
     if [[ $i -eq 0 ]];
       then
@@ -490,7 +548,10 @@ for (( i = 0; i < ${N4_ATROPOS_NUMBER_OF_ITERATIONS}; i++ ))
         N4_WEIGHT_MASK_IMAGES=( ${N4_WEIGHT_MASK_IMAGES[@]} ${POSTERIOR_IMAGE_FILENAMES[${N4_WEIGHT_MASK_POSTERIOR_IDXS[$j]}]} )
       done
 
-    logCmd ${ANTSPATH}ImageMath ${DIMENSION} ${SEGMENTATION_WEIGHT_MASK} PureTissueN4WeightMask ${N4_WEIGHT_MASK_IMAGES[@]}
+    if [[ ${#N4_WEIGHT_MASK_IMAGES[@]} -gt 0 ]];
+      then
+        logCmd ${ANTSPATH}ImageMath ${DIMENSION} ${SEGMENTATION_WEIGHT_MASK} PureTissueN4WeightMask ${N4_WEIGHT_MASK_IMAGES[@]}
+      fi
 
   done
 
@@ -500,7 +561,12 @@ if [[ $KEEP_TMP_IMAGES -eq 0 ]];
   then
     for f in ${TMP_FILES[@]}
       do
-        logCmd rm $f
+        if [[ -e $f ]];
+          then
+            logCmd rm $f
+          else
+            echo "WARNING: expected temp file doesn't exist: $f"
+          fi
       done
   fi
